generate default config file with a test
- split the lib.rs common code into separate files
